+++ packages/node-core/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Updated send_notification PG function to include `_block height` and entity `_id` (#2626)
+
 ### Added
 - Add `TextEncoder` in sandbox, some network package util method is depended on it
 - Added the `--monitor-object-max-depth` flag to mitigate OOM issues when encountering large chunks.(#2644)
@@ -12,9 +15,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed the inconsistency between the `monitor-file-size` flag and the expected behavior.(#2644)
-
-
-### Fixed
 - Improved block range validation in POI endpoint with custom class-validator decorator
 
 ## [16.1.0] - 2024-12-11